run bench with python
